Finished user authentication

+++ frontend/src/components/UserDetails.js
@@ -1,8 +1,10 @@
 import { useState } from "react";
 import { useUsersContext } from "../hooks/useUsersContext"
+import { useAuthContext } from "../hooks/useAuthContext";
 
-const UserDetails = ({user}) => {
+const UserDetails = ({userDet}) => {
     const {dispatch} = useUsersContext();
+    const {user} = useAuthContext();
     const handleClick = async () => {
         setShowPopup(true)
 
@@ -10,10 +12,16 @@ const UserDetails = ({user}) => {
 
     const [showPopup, setShowPopup] = useState(false);
 
-    async function handleDeleteUser() {
+   const handleDeleteUser = async() => {
 
-        const response = await fetch('/api/users/' + user._id, {
-            method:'DELETE'
+        if(!user){
+            return
+        }
+        const response = await fetch('/api/users/' + userDet._id, {
+            method:'DELETE',
+            headers:{
+                'Authorization': `Bearer ${user.token}`
+            }
         })
         const json = await response.json()
         if(response.ok){
@@ -23,12 +31,12 @@ const UserDetails = ({user}) => {
   }
     return(
         <div className="user-details">
-            <h4>{user.hebrewName}</h4>
-            <p>{user.name}<strong> :שם באנגלית </strong></p>
-            <p><strong>תעודת זהות: </strong>{user.id}</p>
-            <p>{user.email}<strong> :אימייל</strong></p>
-            <p><strong>טלפון: </strong>{user.cellphone}</p>
-            <p>{user.role}<strong> :תפקיד</strong></p>
+            <h4>{userDet.hebrewName}</h4>
+            <p>{userDet.name}<strong> :שם באנגלית </strong></p>
+            <p><strong>תעודת זהות: </strong>{userDet.id}</p>
+            <p>{userDet.email}<strong> :אימייל</strong></p>
+            <p><strong>טלפון: </strong>{userDet.cellphone}</p>
+            <p>{userDet.role}<strong> :תפקיד</strong></p>
             <span className="material-symbols-outlined" onClick={handleClick}>delete</span>
             {showPopup && (
             <div className="confirmation-popup">
